requirements.txt
# Database
pymongo==4.6.1

# Web Framework
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# Repository Management
GitPython==3.1.41
requests==2.31.0

# Code Processing
ast-comments==1.0.0
tree-sitter==0.20.4
libcst==1.1.0

# LLM Integration
langchain==0.1.4
langchain-openai==0.0.5
openai==1.10.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2

# Frontend
jinja2==3.1.3
aiofiles==23.2.1 

# Other
python-multipart==0.0.14